requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
# created-by: conda 25.1.0
@EXPLICIT
https://repo.anaconda.com/pkgs/main/osx-arm64/blas-1.0-openblas.conda
https://repo.anaconda.com/pkgs/main/osx-arm64/bzip2-1.0.8-h80987f9_6.conda
https://repo.anaconda.com/pkgs/main/osx-arm64/ca-certificates-2025.2.25-hca03da5_0.conda
https://repo.anaconda.com/pkgs/main/osx-arm64/jpeg-9e-h80987f9_3.conda
https://repo.anaconda.com/pkgs/main/osx-arm64/libcxx-14.0.6-h848a8c0_0.conda
https://repo.anaconda.com/pkgs/main/osx-arm64/libdeflate-1.22-h80987f9_0.conda
https://repo.anaconda.com/pkgs/main/osx-arm64/libffi-3.4.4-hca03da5_1.conda
https://repo.anaconda.com/pkgs/main/osx-arm64/libwebp-base-1.3.2-h80987f9_1.conda
https://repo.anaconda.com/pkgs/main/osx-arm64/llvm-openmp-14.0.6-hc6e5704_0.conda
https://repo.anaconda.com/pkgs/main/osx-arm64/ncurses-6.4-h313beb8_0.conda
https://repo.anaconda.com/pkgs/main/noarch/tzdata-2025a-h04d1e81_0.conda
https://repo.anaconda.com/pkgs/main/osx-arm64/xz-5.6.4-h80987f9_1.conda
https://repo.anaconda.com/pkgs/main/osx-arm64/zlib-1.2.13-h18a0788_1.conda
https://repo.anaconda.com/pkgs/main/osx-arm64/lerc-4.0.0-h313beb8_0.conda
https://repo.anaconda.com/pkgs/main/osx-arm64/libgfortran5-11.3.0-h009349e_28.conda
https://repo.anaconda.com/pkgs/main/osx-arm64/libpng-1.6.39-h80987f9_0.conda
https://repo.anaconda.com/pkgs/main/osx-arm64/lz4-c-1.9.4-h313beb8_1.conda
https://repo.anaconda.com/pkgs/main/osx-arm64/openssl-3.0.16-h02f6b3c_0.conda
https://repo.anaconda.com/pkgs/main/osx-arm64/readline-8.2-h1a28f6b_0.conda
https://repo.anaconda.com/pkgs/main/osx-arm64/tk-8.6.14-h6ba3021_0.conda
https://repo.anaconda.com/pkgs/main/osx-arm64/freetype-2.12.1-h1192e45_0.conda
https://repo.anaconda.com/pkgs/main/osx-arm64/libgfortran-5.0.0-11_3_0_hca03da5_28.conda
https://repo.anaconda.com/pkgs/main/osx-arm64/sqlite-3.45.3-h80987f9_0.conda
https://repo.anaconda.com/pkgs/main/osx-arm64/zstd-1.5.6-hfb09047_0.conda
https://repo.anaconda.com/pkgs/main/osx-arm64/libopenblas-0.3.21-h269037a_0.conda
https://repo.anaconda.com/pkgs/main/osx-arm64/libtiff-4.5.1-hc9ead59_1.conda
https://repo.anaconda.com/pkgs/main/osx-arm64/python-3.10.16-hb885b13_1.conda
https://repo.anaconda.com/pkgs/main/osx-arm64/brotli-python-1.0.9-py310h313beb8_9.conda
https://repo.anaconda.com/pkgs/main/osx-arm64/certifi-2025.1.31-py310hca03da5_0.conda
https://repo.anaconda.com/pkgs/main/noarch/charset-normalizer-3.3.2-pyhd3eb1b0_0.conda
https://repo.anaconda.com/pkgs/main/osx-arm64/click-8.1.7-py310hca03da5_0.conda
https://repo.anaconda.com/pkgs/main/noarch/cycler-0.11.0-pyhd3eb1b0_0.conda
https://repo.anaconda.com/pkgs/main/osx-arm64/idna-3.7-py310hca03da5_0.conda
https://repo.anaconda.com/pkgs/main/osx-arm64/kiwisolver-1.4.8-py310h313beb8_0.conda
https://repo.anaconda.com/pkgs/main/osx-arm64/lcms2-2.16-he93ba84_0.conda
https://repo.anaconda.com/pkgs/main/osx-arm64/numpy-base-2.0.1-py310ha9811e2_1.conda
https://repo.anaconda.com/pkgs/main/osx-arm64/openjpeg-2.5.2-h54b8e55_0.conda
https://repo.anaconda.com/pkgs/main/osx-arm64/packaging-24.2-py310hca03da5_0.conda
https://repo.anaconda.com/pkgs/main/osx-arm64/pyparsing-3.2.0-py310hca03da5_0.conda
https://repo.anaconda.com/pkgs/main/osx-arm64/pysocks-1.7.1-py310hca03da5_0.conda
https://repo.anaconda.com/pkgs/main/noarch/python-tzdata-2023.3-pyhd3eb1b0_0.conda
https://repo.anaconda.com/pkgs/main/osx-arm64/pytz-2024.1-py310hca03da5_0.conda
https://repo.anaconda.com/pkgs/main/osx-arm64/setuptools-75.8.0-py310hca03da5_0.conda
https://repo.anaconda.com/pkgs/main/noarch/six-1.16.0-pyhd3eb1b0_1.conda
https://repo.anaconda.com/pkgs/main/osx-arm64/tornado-6.4.2-py310h80987f9_0.conda
https://repo.anaconda.com/pkgs/main/osx-arm64/unicodedata2-15.1.0-py310h80987f9_1.conda
https://repo.anaconda.com/pkgs/main/osx-arm64/wheel-0.45.1-py310hca03da5_0.conda
https://repo.anaconda.com/pkgs/main/osx-arm64/fonttools-4.55.3-py310h80987f9_0.conda
https://repo.anaconda.com/pkgs/main/osx-arm64/numpy-2.0.1-py310h3b2db8e_1.conda
https://repo.anaconda.com/pkgs/main/osx-arm64/pillow-11.1.0-py310h84e58ab_0.conda
https://repo.anaconda.com/pkgs/main/osx-arm64/pip-25.0-py310hca03da5_0.conda
https://repo.anaconda.com/pkgs/main/osx-arm64/python-dateutil-2.9.0post0-py310hca03da5_2.conda
https://repo.anaconda.com/pkgs/main/osx-arm64/python-dotenv-0.21.0-py310hca03da5_0.conda
https://repo.anaconda.com/pkgs/main/osx-arm64/urllib3-2.3.0-py310hca03da5_0.conda
https://repo.anaconda.com/pkgs/main/osx-arm64/bottleneck-1.4.2-py310hbda83bc_0.conda
https://repo.anaconda.com/pkgs/main/osx-arm64/contourpy-1.3.1-py310h48ca7d4_0.conda
https://repo.anaconda.com/pkgs/main/osx-arm64/numexpr-2.10.1-py310h5d9532f_0.conda
https://repo.anaconda.com/pkgs/main/osx-arm64/requests-2.32.3-py310hca03da5_1.conda
https://repo.anaconda.com/pkgs/main/osx-arm64/matplotlib-base-3.10.0-py310h7ef442a_0.conda
https://repo.anaconda.com/pkgs/main/osx-arm64/pandas-2.2.3-py310hcf29cfe_0.conda
https://repo.anaconda.com/pkgs/main/osx-arm64/matplotlib-3.10.0-py310hca03da5_0.conda

# Core dependencies
flask==3.0.2
python-dotenv==1.0.1
requests==2.31.0
pandas==2.2.0
numpy==1.26.3

# Development dependencies
pytest==8.0.0
black==24.1.1
flake8==7.0.0
